refactor(backend): move agent-runtime into uv workspace at backend/
Restructures the Python service from a standalone tree at
services/agent-runtime/ into a uv workspace rooted at backend/, mirroring
the pattern used by IronTec's nixon backend monorepo. Single uv.lock at
backend/, shared ruff/mypy/pytest config, Dockerfile lives at the
workspace root and copies each member's pyproject before sync to
maximise layer caching.

- services/agent-runtime → backend/agno-agent
- agent_runtime/ Python module → agno_agent/
- pyproject project name agent-runtime → agno-agent
- All `from agent_runtime.*` imports → `from agno_agent.*`
- Dockerfile workspace-aware (COPY per-member pyproject + uv sync
  --no-install-workspace, then COPY source + uv sync)
- CI working-directory backend/, runs tests via
  `uv run --package agno-agent pytest agno-agent/tests`

Leaves room for additional Python services (workers, MCP server, etc) as
sibling members under backend/ without further layout decisions.
