+++ backend/agno-agent/agno_agent/registry.py
@@ -8,10 +8,10 @@ import httpx
 from agno.agent import Agent
 from agno.db.postgres import PostgresDb
 
-from agent_runtime.builder import build_agent
-from agent_runtime.config import settings
-from agent_runtime.db import normalize_pg_url
-from agent_runtime.logging import get_logger
+from agno_agent.builder import build_agent
+from agno_agent.config import settings
+from agno_agent.db import normalize_pg_url
+from agno_agent.logging import get_logger
 
 logger = get_logger(__name__)
