+++ backend/agno-agent/pyproject.toml
@@ -1,0 +1,24 @@
+[project]
+name = "agno-agent"
+version = "0.1.0"
+description = "Zetesis Portal Agent Runtime (Agno-based, shared multi-tenant)"
+requires-python = ">=3.12"
+readme = "README.md"
+license = { text = "MIT" }
+authors = [{ name = "Zetesis Labs" }]
+dependencies = [
+    "agno[anthropic,mcp,openai,os,postgres]>=2.5.16",
+    "fastapi>=0.115",
+    "httpx>=0.28",
+    "psycopg[binary]>=3.2",
+    "pydantic-settings>=2.6",
+    "structlog>=24.1",
+    "uvicorn[standard]>=0.34",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["agno_agent"]
