+++ backend/agno-agent/tests/test_builder.py
@@ -1,13 +1,12 @@
-"""Tests for `agent_runtime.builder.build_model` provider/model mapping."""
+"""Tests for `agno_agent.builder.build_model` provider/model mapping."""
 
 from __future__ import annotations
 
 import pytest
 from agno.models.anthropic import Claude
 from agno.models.openai import OpenAIChat, OpenAIResponses
-
-from agent_runtime.builder import build_model
-from agent_runtime.exceptions import UnsupportedProviderError
+from agno_agent.builder import build_model
+from agno_agent.exceptions import UnsupportedProviderError
 
 
 class TestBuildModel:
