+++ backend/pyproject.toml
@@ -1,0 +1,52 @@
+[project]
+name = "zetesis-backend"
+version = "0.0.0"
+description = "Workspace root for Zetesis Python services."
+requires-python = ">=3.12"
+
+[tool.uv.workspace]
+members = ["agno-agent"]
+
+[tool.uv.sources]
+agno-agent = { workspace = true }
+
+[dependency-groups]
+dev = [
+    "mypy>=1.13",
+    "pyright>=1.1",
+    "pytest-asyncio>=0.24",
+    "pytest>=8.3",
+    "ruff>=0.9",
+]
+
+[tool.ruff]
+line-length = 100
+target-version = "py312"
+
+[tool.ruff.lint]
+select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "S", "SIM", "RUF"]
+ignore = [
+    "E501",
+    "B008",
+]
+
+[tool.ruff.lint.per-file-ignores]
+"**/tests/**/*.py" = ["S101"]
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+
+[tool.mypy]
+python_version = "3.12"
+strict = true
+warn_return_any = true
+warn_unused_configs = true
+ignore_missing_imports = false
+
+[[tool.mypy.overrides]]
+module = ["agno.*"]
+ignore_missing_imports = true
+
+[tool.pytest.ini_options]
+asyncio_mode = "auto"
